+++ pypeit/debugger.py
@@ -1,3 +1,6 @@
+"""
+Module to setup the PypeIt debugger
+"""
 from __future__ import (print_function, absolute_import, division, unicode_literals)
 
 import matplotlib.pyplot as plt
@@ -5,11 +8,11 @@ import numpy as np
 
 # These need to be outside of the def's
 try:
-    from pypit.ginga import show_image
+    from pypeit.ginga import show_image
 except ImportError:  # Ginga is not yet required
     pass
 else:
-    from pypit.ginga import chk_arc_tilts, clear_canvas
+    from pypeit.ginga import chk_arc_tilts, clear_canvas
 # Moved to the top and changed to only import set_trace
 from pdb import set_trace
