+++ pypeit/deprecated/arspecobj.py
