+++ pypeit/telescopes.py
@@ -8,7 +8,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from pypit.par.pypitpar import TelescopePar
+from pypeit.par.pypeitpar import TelescopePar
 
 #TODO: Remove 'Par' from class name?
